moved settings-related code from `CppCheckExecutor` to `CmdLineParser`
