+++ cli/cppcheckexecutor.cpp
@@ -18,7 +18,6 @@
 
 #include "cppcheckexecutor.h"
 
-#include "addoninfo.h"
 #include "analyzerinfo.h"
 #include "checkersreport.h"
 #include "cmdlinelogger.h"
@@ -28,11 +27,7 @@
 #include "cppcheck.h"
 #include "errorlogger.h"
 #include "errortypes.h"
-#include "filelister.h"
 #include "filesettings.h"
-#include "library.h"
-#include "path.h"
-#include "pathmatch.h"
 #include "settings.h"
 #include "singleexecutor.h"
 #include "suppressions.h"
@@ -47,16 +42,13 @@
 #endif
 
 #include <algorithm>
-#include <cassert>
 #include <cstdio>
 #include <cstdlib> // EXIT_SUCCESS and EXIT_FAILURE
 #include <ctime>
 #include <iostream>
-#include <iterator>
 #include <list>
 #include <set>
 #include <sstream> // IWYU pragma: keep
-#include <unordered_set>
 #include <utility>
 #include <vector>
 
@@ -72,44 +64,26 @@
 #include <windows.h>
 #endif
 
-namespace {
-    class XMLErrorMessagesLogger : public ErrorLogger
+class CmdLineLoggerStd : public CmdLineLogger
+{
+public:
+    CmdLineLoggerStd() = default;
+
+    void printMessage(const std::string &message) override
     {
-        void reportOut(const std::string & outmsg, Color /*c*/ = Color::Reset) override
-        {
-            std::cout << outmsg << std::endl;
-        }
+        printRaw("cppcheck: " + message);
+    }
 
-        void reportErr(const ErrorMessage &msg) override
-        {
-            reportOut(msg.toXML());
-        }
-
-        void reportProgress(const std::string & /*filename*/, const char /*stage*/[], const std::size_t /*value*/) override
-        {}
-    };
-
-    class CmdLineLoggerStd : public CmdLineLogger
+    void printError(const std::string &message) override
     {
-    public:
-        CmdLineLoggerStd() = default;
+        printMessage("error: " + message);
+    }
 
-        void printMessage(const std::string &message) override
-        {
-            printRaw("cppcheck: " + message);
-        }
-
-        void printError(const std::string &message) override
-        {
-            printMessage("error: " + message);
-        }
-
-        void printRaw(const std::string &message) override
-        {
-            std::cout << message << std::endl;
-        }
-    };
-}
+    void printRaw(const std::string &message) override
+    {
+        std::cout << message << std::endl;
+    }
+};
 
 class CppCheckExecutor::StdLogger : public ErrorLogger
 {
@@ -195,172 +169,22 @@ private:
 /*static*/ FILE* CppCheckExecutor::mExceptionOutput = stdout;
 #endif
 
-bool CppCheckExecutor::parseFromArgs(Settings &settings, int argc, const char* const argv[])
-{
-    CmdLineLoggerStd logger;
-    CmdLineParser parser(logger, settings, settings.nomsg, settings.nofail);
-    const bool success = parser.parseFromArgs(argc, argv);
-
-    if (success) {
-        if (parser.getShowVersion() && !parser.getShowErrorMessages()) {
-            if (!settings.cppcheckCfgProductName.empty()) {
-                logger.printRaw(settings.cppcheckCfgProductName);
-            } else {
-                const char * const extraVersion = CppCheck::extraVersion();
-                if (*extraVersion != 0)
-                    logger.printRaw(std::string("Cppcheck ") + CppCheck::version() + " ("+ extraVersion + ')');
-                else
-                    logger.printRaw(std::string("Cppcheck ") + CppCheck::version());
-            }
-        }
-
-        if (parser.getShowErrorMessages()) {
-            XMLErrorMessagesLogger xmlLogger;
-            std::cout << ErrorMessage::getXMLHeader(settings.cppcheckCfgProductName);
-            CppCheck::getErrorMessages(xmlLogger);
-            std::cout << ErrorMessage::getXMLFooter() << std::endl;
-        }
-
-        if (parser.exitAfterPrinting()) {
-            Settings::terminate();
-            return true;
-        }
-    } else {
-        return false;
-    }
-
-    // Libraries must be loaded before FileLister is executed to ensure markup files will be
-    // listed properly.
-    if (!loadLibraries(settings))
-        return false;
-
-    if (!loadAddons(settings))
-        return false;
-
-    // Check that all include paths exist
-    {
-        for (std::list<std::string>::iterator iter = settings.includePaths.begin();
-             iter != settings.includePaths.end();
-             ) {
-            const std::string path(Path::toNativeSeparators(*iter));
-            if (Path::isDirectory(path))
-                ++iter;
-            else {
-                // TODO: this bypasses the template format and other settings
-                // If the include path is not found, warn user and remove the non-existing path from the list.
-                if (settings.severity.isEnabled(Severity::information))
-                    std::cout << "(information) Couldn't find path given by -I '" << path << '\'' << std::endl;
-                iter = settings.includePaths.erase(iter);
-            }
-        }
-    }
-
-    // Output a warning for the user if he tries to exclude headers
-    const std::vector<std::string>& ignored = parser.getIgnoredPaths();
-    const bool warn = std::any_of(ignored.cbegin(), ignored.cend(), [](const std::string& i) {
-        return Path::isHeader(i);
-    });
-    if (warn) {
-        logger.printMessage("filename exclusion does not apply to header (.h and .hpp) files.");
-        logger.printMessage("Please use --suppress for ignoring results from the header files.");
-    }
-
-    const std::vector<std::string>& pathnamesRef = parser.getPathNames();
-    const std::list<FileSettings>& fileSettingsRef = parser.getFileSettings();
-
-    // the inputs can only be used exclusively - CmdLineParser should already handle this
-    assert(!(!pathnamesRef.empty() && !fileSettingsRef.empty()));
-
-    if (!fileSettingsRef.empty()) {
-        std::list<FileSettings> fileSettings;
-        if (!settings.fileFilters.empty()) {
-            // filter only for the selected filenames from all project files
-            std::copy_if(fileSettingsRef.cbegin(), fileSettingsRef.cend(), std::back_inserter(fileSettings), [&](const FileSettings &fs) {
-                return matchglobs(settings.fileFilters, fs.filename);
-            });
-            if (fileSettings.empty()) {
-                logger.printError("could not find any files matching the filter.");
-                return false;
-            }
-        }
-        else {
-            fileSettings = fileSettingsRef;
-        }
-
-        // sort the markup last
-        std::copy_if(fileSettings.cbegin(), fileSettings.cend(), std::back_inserter(mFileSettings), [&](const FileSettings &fs) {
-            return !settings.library.markupFile(fs.filename) || !settings.library.processMarkupAfterCode(fs.filename);
-        });
-
-        std::copy_if(fileSettings.cbegin(), fileSettings.cend(), std::back_inserter(mFileSettings), [&](const FileSettings &fs) {
-            return settings.library.markupFile(fs.filename) && settings.library.processMarkupAfterCode(fs.filename);
-        });
-    }
-
-    if (!pathnamesRef.empty()) {
-        std::list<std::pair<std::string, std::size_t>> filesResolved;
-        // TODO: this needs to be inlined into PathMatch as it depends on the underlying filesystem
-#if defined(_WIN32)
-        // For Windows we want case-insensitive path matching
-        const bool caseSensitive = false;
-#else
-        const bool caseSensitive = true;
-#endif
-        // Execute recursiveAddFiles() to each given file parameter
-        const PathMatch matcher(ignored, caseSensitive);
-        for (const std::string &pathname : pathnamesRef) {
-            const std::string err = FileLister::recursiveAddFiles(filesResolved, Path::toNativeSeparators(pathname), settings.library.markupExtensions(), matcher);
-            if (!err.empty()) {
-                // TODO: bail out?
-                logger.printMessage(err);
-            }
-        }
-
-        std::list<std::pair<std::string, std::size_t>> files;
-        if (!settings.fileFilters.empty()) {
-            std::copy_if(filesResolved.cbegin(), filesResolved.cend(), std::inserter(files, files.end()), [&](const decltype(filesResolved)::value_type& entry) {
-                return matchglobs(settings.fileFilters, entry.first);
-            });
-            if (files.empty()) {
-                logger.printError("could not find any files matching the filter.");
-                return false;
-            }
-        }
-        else {
-            files = std::move(filesResolved);
-        }
-
-        // sort the markup last
-        std::copy_if(files.cbegin(), files.cend(), std::inserter(mFiles, mFiles.end()), [&](const decltype(files)::value_type& entry) {
-            return !settings.library.markupFile(entry.first) || !settings.library.processMarkupAfterCode(entry.first);
-        });
-
-        std::copy_if(files.cbegin(), files.cend(), std::inserter(mFiles, mFiles.end()), [&](const decltype(files)::value_type& entry) {
-            return settings.library.markupFile(entry.first) && settings.library.processMarkupAfterCode(entry.first);
-        });
-    }
-
-    if (mFiles.empty() && mFileSettings.empty()) {
-        logger.printError("could not find or open any of the paths given.");
-        if (!ignored.empty())
-            logger.printMessage("Maybe all paths were ignored?");
-        return false;
-    }
-
-    return true;
-}
-
 int CppCheckExecutor::check(int argc, const char* const argv[])
 {
     CheckUnusedFunctions::clear();
 
     Settings settings;
-    if (!parseFromArgs(settings, argc, argv)) {
+    CmdLineLoggerStd logger;
+    CmdLineParser parser(logger, settings, settings.nomsg, settings.nofail);
+    if (!parser.fillSettingsFromArgs(argc, argv)) {
         return EXIT_FAILURE;
     }
     if (Settings::terminated()) {
         return EXIT_SUCCESS;
     }
+
+    mFiles = parser.getFiles();
+    mFileSettings = parser.getFileSettings();
 
     mStdLogger = new StdLogger(settings);
     CppCheck cppCheck(*mStdLogger, true, executeCommand);
@@ -490,49 +314,6 @@ void CppCheckExecutor::StdLogger::writeCheckersReport() const
 
 }
 
-bool CppCheckExecutor::loadLibraries(Settings& settings)
-{
-    if (!tryLoadLibrary(settings.library, settings.exename, "std.cfg")) {
-        const std::string msg("Failed to load std.cfg. Your Cppcheck installation is broken, please re-install.");
-#ifdef FILESDIR
-        const std::string details("The Cppcheck binary was compiled with FILESDIR set to \""
-                                  FILESDIR "\" and will therefore search for "
-                                  "std.cfg in " FILESDIR "/cfg.");
-#else
-        const std::string cfgfolder(Path::fromNativeSeparators(Path::getPathFromFilename(settings.exename)) + "cfg");
-        const std::string details("The Cppcheck binary was compiled without FILESDIR set. Either the "
-                                  "std.cfg should be available in " + cfgfolder + " or the FILESDIR "
-                                  "should be configured.");
-#endif
-        std::cout << msg << " " << details << std::endl;
-        return false;
-    }
-
-    bool result = true;
-    for (const auto& lib : settings.libraries) {
-        if (!tryLoadLibrary(settings.library, settings.exename, lib.c_str())) {
-            result = false;
-        }
-    }
-    return result;
-}
-
-bool CppCheckExecutor::loadAddons(Settings& settings)
-{
-    bool result = true;
-    for (const std::string &addon: settings.addons) {
-        AddonInfo addonInfo;
-        const std::string failedToGetAddonInfo = addonInfo.getAddonInfo(addon, settings.exename);
-        if (!failedToGetAddonInfo.empty()) {
-            std::cout << failedToGetAddonInfo << std::endl;
-            result = false;
-            continue;
-        }
-        settings.addonInfos.emplace_back(std::move(addonInfo));
-    }
-    return result;
-}
-
 #ifdef _WIN32
 // fix trac ticket #439 'Cppcheck reports wrong filename for filenames containing 8-bit ASCII'
 static inline std::string ansiToOEM(const std::string &msg, bool doConvert)
@@ -634,50 +415,6 @@ FILE* CppCheckExecutor::getExceptionOutput()
     return mExceptionOutput;
 }
 #endif
-
-bool CppCheckExecutor::tryLoadLibrary(Library& destination, const std::string& basepath, const char* filename)
-{
-    const Library::Error err = destination.load(basepath.c_str(), filename);
-
-    if (err.errorcode == Library::ErrorCode::UNKNOWN_ELEMENT)
-        std::cout << "cppcheck: Found unknown elements in configuration file '" << filename << "': " << err.reason << std::endl;
-    else if (err.errorcode != Library::ErrorCode::OK) {
-        std::cout << "cppcheck: Failed to load library configuration file '" << filename << "'. ";
-        switch (err.errorcode) {
-        case Library::ErrorCode::OK:
-            break;
-        case Library::ErrorCode::FILE_NOT_FOUND:
-            std::cout << "File not found";
-            break;
-        case Library::ErrorCode::BAD_XML:
-            std::cout << "Bad XML";
-            break;
-        case Library::ErrorCode::UNKNOWN_ELEMENT:
-            std::cout << "Unexpected element";
-            break;
-        case Library::ErrorCode::MISSING_ATTRIBUTE:
-            std::cout << "Missing attribute";
-            break;
-        case Library::ErrorCode::BAD_ATTRIBUTE_VALUE:
-            std::cout << "Bad attribute value";
-            break;
-        case Library::ErrorCode::UNSUPPORTED_FORMAT:
-            std::cout << "File is of unsupported format version";
-            break;
-        case Library::ErrorCode::DUPLICATE_PLATFORM_TYPE:
-            std::cout << "Duplicate platform type";
-            break;
-        case Library::ErrorCode::PLATFORM_TYPE_REDEFINED:
-            std::cout << "Platform type redefined";
-            break;
-        }
-        if (!err.reason.empty())
-            std::cout << " '" + err.reason + "'";
-        std::cout << std::endl;
-        return false;
-    }
-    return true;
-}
 
 /**
  * Execute a shell command and read the output from it. Returns true if command terminated successfully.
